# Conflicts:
#	2024_winapigamep_framework/2024_winapigamep_framework.vcxproj.filters
#	Output/build_debug/2024_winapigamep_framework.exe

+++ 2024_winapigamep_framework/KingState.h
@@ -19,6 +19,6 @@ private:
     bool isAttack;
     bool isEnd;
 
-    int currentAttackCount = 0;
+    float currentAttackCount = 0;
 };
